frontend-app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-cloud-texttospeech==2.14.1
google-generativeai==0.3.1
python-multipart==0.0.6
pydantic==2.4.2
requests==2.31.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
sqlalchemy==2.0.23
aiofiles==23.2.1
beautifulsoup4==4.12.2
lxml==4.9.3 
pydub==0.25.1
inflect
google-cloud-translate
